feat(stress-test): custom transactions to tx generator, result summary (#880)
Description
---
Allow a custom manifest to be used with transaction generator
Fetch results and summarize stress test results 
Allow fee instructions to be generated from transaction manifests
Add test bench app for tariswaps
Fix bug in autofiller bug where input refs are "upgraded" to inputs 

Motivation and Context
---
Support custom stress test transactions. 

Manifests can now specify fee instructions by optionally declaring a
`fee_main()` function

Adds test runner app that sets up and submits thousands of swaps.

How Has This Been Tested?
---
Manually, tests updated as required

What process can a PR reviewer use to test or verify this change?
---
Use transaction generator with custom manifest

```
cargo run --package transaction_generator --bin transaction_generator --release -- write -n 1000000 -o ~/Litterbox/transactions.bin --overwrite -m manifests/tariswap_do_swap.rs -gswap_component=component_7e476f070b8926bbee20874ee1ecbe98026ac03c8f96b4a896828677ba13afa4 -gamt=10 -gin_account=component_afa270d4b522db62a72434f2d86c30cfa590c2ea8ac830731dca2b7b3cf2ef7a -gout_account=component_afa270d4b522db62a72434f2d86c30cfa590c2ea8ac830731dca2b7b3cf2ef7a -gtoken_a=resource_0101010101010101010101010101010101010101010101010101010101010101 -gtoken_b=resource_1ef527d83b8a3e4b3010e5fcacdf97bff60cbf33ebeec09d45772c79569e8efb  -k=f5e213e87b8630c269b140d7564dfb7b77a02f166f15be4eacfae968f4859c0a
```

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify

+++ applications/tari_dan_wallet_daemon/src/main.rs
@@ -23,8 +23,10 @@
 use std::{fs, panic, process};
 
 use tari_common::initialize_logging;
+use tari_crypto::{keys::PublicKey, ristretto::RistrettoPublicKey};
 use tari_dan_app_utilities::configuration::load_configuration;
-use tari_dan_wallet_daemon::{cli::Cli, config::ApplicationConfig, run_tari_dan_wallet_daemon};
+use tari_dan_wallet_daemon::{cli::Cli, config::ApplicationConfig, initialize_wallet_sdk, run_tari_dan_wallet_daemon};
+use tari_dan_wallet_sdk::apis::key_manager;
 use tari_shutdown::Shutdown;
 
 #[tokio::main]
@@ -38,9 +40,20 @@ async fn main() -> Result<(), anyhow::Error> {
     }));
 
     let cli = Cli::init();
+
     let config_path = cli.common.config_path();
     let cfg = load_configuration(config_path, true, &cli)?;
     let config = ApplicationConfig::load_from(&cfg)?;
+
+    if let Some(index) = cli.derive_secret {
+        let sdk = initialize_wallet_sdk(&config)?;
+        let secret = sdk
+            .key_manager_api()
+            .derive_key(key_manager::TRANSACTION_BRANCH, index)?;
+        println!("Secret: {}", secret.key.reveal());
+        println!("Public key: {}", RistrettoPublicKey::from_secret_key(&secret.key));
+        return Ok(());
+    }
 
     // Remove the file if it was left behind by a previous run
     let _file = fs::remove_file(config.common.base_path.join("pid"));
